requirements.txt
openai-whisper
anthropic
elevenlabs
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
watchdog
pydub
flask
gunicorn
